all changes done as requested in comments of PR #56
