+++ include/engine/dataman/DataManReader.h
@@ -19,7 +19,7 @@
 // supported capsules
 #include "capsule/heap/STLVector.h"
 
-#include "utilities/realtime/dataman/DataManager.h"
+#include "utilities/realtime/dataman/DataMan.h"
 
 namespace adios
 {
@@ -38,6 +38,7 @@ public:
      * @param debugMode
      * @param nthreads
      */
+    using json = nlohmann::json;
     DataManReader(ADIOS &adios, const std::string &name,
                   const std::string accessMode, MPI_Comm mpiComm,
                   const Method &method);
@@ -108,7 +109,7 @@ private:
     /// m_Transports
 
     bool m_DoRealTime = false;
-    realtime::DataManager m_Man;
+    realtime::DataMan m_Man;
     std::function<void(const void *, std::string, std::string, std::string,
                        Dims)>
         m_CallBack; ///< call back function
